Fix incorrect copyright headers
The qTox Project is not associated with the Tox Project in any way, with the
exception of qTox using the Tox Projet's toxcore collection of libraries.
In particular, the Tox Projet does not own copyright over the qTox Project's
qTox collection of software, source code, and assets.
The qTox Project's assets are under the sole copyright of the qTox
contributors, and no partiular rights are granted to the Tox Project.