User manual wording+changes
Misleading information

The profile is actually in the top left corner, not the top right.

Updated and reworded the profile section

The section was given a unified look and some options were reworded to either be more clear or to fix grammatical errors.

Minor rewording of the general settings section

No capital S in status

Updated and reworded the chat section

Updated and reworded the theme section

simleys->smileys
Also they are not being replaced with unicode emoticons but instead with actual images, rectified.

Updated and reworded the connection settings section

Minor rewording of the NoSpam section

Minor rewording of the audio+video section

Minor rewording of the advanced section

Updated and reworded the groupchats section

Updated and reworded the multi window mode section

normaly->normally

if profile->if the profile